Adds gltf loader destruction in model destroy function
