+++ Specs/Scene/ModelExperimental/ModelExperimentalSceneGraphSpec.js
@@ -1,21 +1,15 @@
 import {
   Axis,
-  Cartesian3,
-  defaultValue,
-  HeadingPitchRange,
   Matrix4,
-  ModelExperimental,
   ModelExperimentalSceneGraph,
-  when,
+  ResourceCache,
 } from "../../../Source/Cesium.js";
 import createScene from "../../createScene.js";
-import pollToPromise from "../../pollToPromise.js";
+import loadAndZoomToModelExperimental from "./loadModelExperimentalForSpec.js";
 
 describe(
   "Scene/ModelExperimental/ModelExperimentalSceneGraph",
   function () {
-    var boxTexturedGlbUrl =
-      "./Data/Models/GltfLoader/BoxTextured/glTF-Binary/BoxTextured.glb";
     var parentGltfUrl = "./Data/Cesium3DTiles/GltfContent/glTF/parent.gltf";
 
     var scene;
@@ -28,74 +22,46 @@ describe(
       scene.destroyForSpecs();
     });
 
-    function loadModel(options) {
-      var model = ModelExperimental.fromGltf({
-        url: options.url,
-        upAxis: options.upAxis,
-        forwardAxis: options.forwardAxis,
-      });
-      scene.primitives.add(model);
-      zoomTo(model);
-
-      return pollToPromise(
-        function () {
-          scene.renderForSpecs();
-          return model.ready;
-        },
-        { timeout: 10000 }
-      )
-        .then(function () {
-          return model;
-        })
-        .otherwise(function () {
-          return when.reject(model);
-        });
-    }
-
-    function zoomTo(model) {
-      model.zoomTo = function (zoom) {
-        zoom = defaultValue(zoom, 4.0);
-
-        var camera = scene.camera;
-        var center = Matrix4.multiplyByPoint(
-          model.modelMatrix,
-          model.boundingSphere.center,
-          new Cartesian3()
-        );
-        var r =
-          zoom * Math.max(model.boundingSphere.radius, camera.frustum.near);
-        camera.lookAt(center, new HeadingPitchRange(0.0, 0.0, r));
-      };
-    }
+    afterEach(function () {
+      scene.primitives.removeAll();
+      ResourceCache.clearForSpecs();
+    });
 
     it("creates scene nodes from a model", function () {
-      return loadModel({ url: boxTexturedGlbUrl }).then(function (model) {
-        var sceneGraph = model._sceneGraph;
-        var modelComponents = sceneGraph._modelComponents;
-        expect(sceneGraph).toBeDefined();
-        expect(sceneGraph._sceneNodes.length).toEqual(
-          modelComponents.nodes.length
-        );
-      });
+      return loadAndZoomToModelExperimental({ url: parentGltfUrl }, scene).then(
+        function (model) {
+          var sceneGraph = model._sceneGraph;
+          var modelComponents = sceneGraph._modelComponents;
+          expect(sceneGraph).toBeDefined();
+          expect(sceneGraph._sceneNodes.length).toEqual(
+            modelComponents.nodes.length
+          );
+        }
+      );
     });
 
     it("traverses scene graph correctly", function () {
-      return loadModel({ url: boxTexturedGlbUrl }).then(function (model) {
-        var sceneGraph = model._sceneGraph;
-        var modelComponents = sceneGraph._modelComponents;
-        var sceneNodes = sceneGraph._sceneNodes;
+      return loadAndZoomToModelExperimental({ url: parentGltfUrl }, scene).then(
+        function (model) {
+          var sceneGraph = model._sceneGraph;
+          var modelComponents = sceneGraph._modelComponents;
+          var sceneNodes = sceneGraph._sceneNodes;
 
-        expect(sceneNodes[1].node).toEqual(modelComponents.nodes[0]);
-        expect(sceneNodes[0].node).toEqual(modelComponents.nodes[1]);
-      });
+          expect(sceneNodes[1].node).toEqual(modelComponents.nodes[0]);
+          expect(sceneNodes[0].node).toEqual(modelComponents.nodes[1]);
+        }
+      );
     });
 
     it("propagates node transforms correctly", function () {
-      return loadModel({
-        url: parentGltfUrl,
-        upAxis: Axis.Z,
-        forwardAxis: Axis.X,
-      }).then(function (model) {
+      return loadAndZoomToModelExperimental(
+        {
+          url: parentGltfUrl,
+          upAxis: Axis.Z,
+          forwardAxis: Axis.X,
+        },
+        scene
+      ).then(function (model) {
         var sceneGraph = model._sceneGraph;
         var modelComponents = sceneGraph._modelComponents;
         var sceneNodes = sceneGraph._sceneNodes;
